Move all profile utility related classes to own package
