+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/conv14_50/datatypes14_50/ElementDefinition14_50.java
@@ -17,7 +17,7 @@ import org.hl7.fhir.convertors.conv14_50.datatypes14_50.primitivetypes14_50.Uri1
 import org.hl7.fhir.convertors.conv14_50.resources14_50.Enumerations14_50;
 import org.hl7.fhir.dstu2016may.model.ElementDefinition;
 import org.hl7.fhir.exceptions.FHIRException;
-import org.hl7.fhir.r5.conformance.ProfileUtilities;
+import org.hl7.fhir.r5.conformance.profile.ProfileUtilities;
 import org.hl7.fhir.utilities.Utilities;
 
 public class ElementDefinition14_50 {
